Do not not sleep but instead block until messages are availabe
As the default sleep duration is one second, this starts message
processing on avarage 500ms sooner in use cases with often empty queues.

As messages with a dead worker are only picked up during queue
maintenance, it is ensured that queue maintenance happens at least every
10 seconds.

The timeout until a worker is considered dead is reduced to 10 seconds.
This way a message is retried latest after 20 seconds if a worker died.